Splits Request and Response Xsd validation As a convenience,
