+++ Digipost.Signature.Api.Client.Core/Internal/XsdResponseValidationHandler.cs
@@ -1,0 +1,22 @@
+﻿using System.Linq;
+using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
+using Digipost.Signature.Api.Client.Core.Exceptions;
+using Digipost.Signature.Api.Client.Core.Xsd;
+
+namespace Digipost.Signature.Api.Client.Core.Internal
+{
+    internal class XsdResponseValidationHandler : XsdValidationHandler
+    {
+
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        {
+            var result = await base.SendAsync(request, cancellationToken);
+
+            await ValidateXmlAndThrowIfInvalid(result.Content, cancellationToken);
+
+            return result;
+        }
+    }
+}
