+++ Digipost.Signature.Api.Client.Core/Internal/XsdValidationHandler.cs
@@ -1,35 +1,15 @@
-﻿using System;
 using System.Linq;
 using System.Net.Http;
-using System.Net.Mime;
-using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Xml;
-using Difi.Felles.Utility;
 using Digipost.Signature.Api.Client.Core.Exceptions;
 using Digipost.Signature.Api.Client.Core.Xsd;
-using log4net;
 
 namespace Digipost.Signature.Api.Client.Core.Internal
 {
     public class XsdValidationHandler : DelegatingHandler
     {
-        private static readonly ILog Log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
-
-
-        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-        {
-            await ValidateXmlAndThrowIfInvalid(request.Content, cancellationToken);
-            
-            var result = await base.SendAsync(request, cancellationToken);
-
-            await ValidateXmlAndThrowIfInvalid(result.Content, cancellationToken);
-
-            return result;
-        }
-
-        private static async Task ValidateXmlAndThrowIfInvalid(HttpContent content, CancellationToken cancellationToken)
+        protected static async Task ValidateXmlAndThrowIfInvalid(HttpContent content, CancellationToken cancellationToken)
         {
             if (content?.Headers.ContentType?.MediaType == "multipart/mixed")
             {
